pyproject.toml
[tool.poetry]
name = "inventory-manager"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "inventory_manager"}]

[tool.poetry.dependencies]
python = "^3.9"
django = "4.0.1"
reactivated = "^0.30.2"
djangorestframework = "^3.14.0"
django-widget-tweaks = "^1.4.12"
checkdigit = "^0.5.0"
pytz = "^2023.3"
psycopg2-binary = "^2.9.6"
django-cors-headers = "^4.1.0"
python-barcode = "^0.14.0"
pillow = "^9.5.0"
django-cleanup = "^8.0.0"
django-rq = "^2.8.1"
redis = "^4.5.5"
natsort = "^8.3.1"
shortuuid = "^1.0.11"
django-stubs = {extras = ["compatible-mypy"], version = "^4.2.3"}
djangorestframework-stubs = {extras = ["compatible-mypy"], version = "^3.14.2"}
uvicorn = "^0.22.0"
cattrs = "^23.1.2"
django-typomatic = "^2.3.2"
requests-html = "^0.10.0"
beautifulsoup4 = "^4.12.2"
sqlparse = "^0.4.4"
sql-metadata = "^2.8.0"


[tool.poetry.group.dev.dependencies]
isort = "5.10.1"
flake8 = "^6.0.0"
mypy = "^1.3.0"
djangorestframework-stubs = "^3.14.1"
types-pillow = "^10.0.0.2"
types-redis = "^4.6.0.3"
types-beautifulsoup4 = "^4.12.0.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
